refactor: landing 화면 이름 변경 및 폴더 구조 변경, 스타일 요소 한 파일로 분리
