+++ src/Page/Landing/index.ts
@@ -1,0 +1,3 @@
+export { default as AgreementPage } from "./AgreementPage";
+export { default as LandingMainPage } from "./LandingMainPage";
+export { default as SignUpPage } from "./SignUpPage";
